Rename file voter -> experiment
